requirements.txt
flask
playwright
python-dotenv